moved error alert actions to root but interfaces to types
