+++ src/app/static/src/app/store/root/actions.ts
@@ -2,13 +2,20 @@ import {ActionContext, ActionTree} from "vuex";
 import {RootState} from "../../root";
 import {StepDescription} from "../stepper/stepper";
 import {RootMutation} from "./mutations";
+import {ErrorsMutation} from "../errors/mutations";
 import {LanguageActions} from "../language/language";
 import {changeLanguage} from "../language/actions";
 import i18next from "i18next";
+import {api} from "../../apiService";
+import {ErrorReportManualDetails} from "../../types";
+import {VersionInfo} from "../../generated";
+import {currentHintVersion} from "../../hintVersion";
 
 
 export interface RootActions extends LanguageActions<RootState> {
     validate: (store: ActionContext<RootState, RootState>) => void;
+    generateErrorReport: (store: ActionContext<RootState, RootState>,
+        payload: ErrorReportManualDetails) => void;
 }
 
 export const actions: ActionTree<RootState, RootState> & RootActions = {
@@ -77,5 +84,36 @@ export const actions: ActionTree<RootState, RootState> & RootActions = {
 
         await Promise.all(actions);
         commit({type: RootMutation.SetUpdatingLanguage, payload: false});
+    },
+
+    async generateErrorReport(context, payload) {
+        const {dispatch, rootState, getters, commit} = context
+        const data = {
+            email: payload.email || rootState.currentUser,
+            country: rootState.baseline.country || "no associated country",
+            projectName: rootState.projects.currentProject?.name || "no associated project",
+            browserAgent: navigator.userAgent,
+            timeStamp: new Date().toISOString(),
+            jobId: rootState.modelRun.modelRunId || "no associated jobId",
+            description: payload.description,
+            section: payload.section,
+            stepsToReproduce: payload.stepsToReproduce,
+            versions: {hint: currentHintVersion, ...rootState.hintrVersion.hintrVersion as VersionInfo},
+            errors: getters.errors
+        }
+
+        await api<ErrorsMutation, ErrorsMutation>(context)
+            .withSuccess(`errors/${ErrorsMutation.ErrorReportSuccess}` as ErrorsMutation, true)
+            .withError(`errors/${ErrorsMutation.ErrorReportError}` as ErrorsMutation, true)
+            .postAndReturn("error-report", data)
+            .then(() => {
+                if (rootState.projects.currentProject && !rootState.errors.errorReportError) {
+                    dispatch("projects/cloneProject",
+                        {
+                            emails: ["naomi-support@imperial.ac.uk"],
+                            projectId: rootState.projects.currentProject!.id
+                        })
+                }
+            })
     }
 };
